1. Метод добавления текста с длинной периметра полигона доработан и перенесен в CustomPolygon - текст появляется/исчезает при клике. 2. В итоге подсчёт периметра полигона будет через polygon.distance.

+++ app/src/main/java/com/veygard/frontiermap/domain/models/GeoCluster.kt
@@ -1,3 +1,3 @@
 package com.veygard.frontiermap.domain.models
 
-data  class GeoCluster(val list: List<MultiPolygon>, val perimeterLengthKm: Float)
+data  class GeoCluster(val list: List<MultiPolygon>, val perimeterLengthKm: Int)
